feat: wire /v2/exporting, /cluster/v2/exporting and /actuator/exporting to dynamic config
ExportingServices, ClusterExportingServices, and the actuator's
ExportingEndpoint all went through ExportingRequestBroadcaster, which
sends the same raw admin wire requests PR #60541's rejection just
disabled - so as of that change, none of the three endpoints backing
exporting pause/resume/status actually worked. This wires all three
onto DynamicConfigExportingStateController instead:

- ExportingServices takes ExportingStateController.ByTenant, resolved
  once per physical tenant at construction, matching the tenant it was
  already scoped to.
- ClusterExportingServices takes ExportingStateController.ClusterWide
  and is now a thin ServiceException-mapping wrapper; the per-tenant
  fan-out and status folding it used to do itself moved into
  DynamicConfigExportingStateController, which already has its own
  coverage for that logic (ADR 003 D2, D4).
- ExportingEndpoint keeps its existing per-tenant fan-out over
  ExportingStateController.getByTenant(id), preserving today's
  "every tenant attempted, first failure decides the response"
  contract.

The response contract is unchanged: the new ExportingStatus
(dynamic-config module) has the same four values as the retired
gateway.admin.ExportingStatus it replaces, so ExportingResponseMapper
only needed an import swap.

ExportingRequestBroadcaster and the old ExportingStatus are now dead
and removed. ClusterRocksDbMigrationStatusProvider reused
ExportingRequestBroadcaster.validateTopology for an unrelated
purpose (skipping a partition should never look like it was
fully migrated); that method moves to a new TopologyValidation
utility so it survives independently of the exporting code path.
