--- zeebe/gateway/src/test/java/io/camunda/zeebe/gateway/admin/ExportingRequestBroadcasterTest.java
@@ -1,396 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.zeebe.gateway.admin;
-
-import static io.camunda.cluster.PhysicalTenantIds.DEFAULT_PHYSICAL_TENANT_ID;
-import static io.camunda.zeebe.gateway.admin.ExportingRequestBroadcasterTest.RequestMatcher.requestTo;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
-import static org.junit.jupiter.api.Named.named;
-import static org.junit.jupiter.params.provider.Arguments.arguments;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.argThat;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-import io.atomix.cluster.BrokerMemberId;
-import io.camunda.zeebe.broker.client.api.BrokerClient;
-import io.camunda.zeebe.broker.client.api.BrokerClusterState;
-import io.camunda.zeebe.broker.client.api.BrokerTopologyManager;
-import io.camunda.zeebe.broker.client.api.dto.BrokerRequest;
-import io.camunda.zeebe.broker.client.api.dto.BrokerResponse;
-import io.camunda.zeebe.dynamic.config.state.PartitionState.State;
-import io.camunda.zeebe.protocol.impl.encoding.AdminResponse;
-import io.camunda.zeebe.protocol.record.PartitionHealthStatus;
-import java.nio.charset.StandardCharsets;
-import java.time.Duration;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import org.jspecify.annotations.NonNull;
-import org.jspecify.annotations.Nullable;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
-import org.mockito.ArgumentMatcher;
-
-public class ExportingRequestBroadcasterTest {
-
-  @ParameterizedTest
-  @MethodSource("validTopologies")
-  void shouldPauseOnAllBrokersAndPartitions(final BrokerClusterState topology) {
-    // given
-    final var client = setupBrokerClient(topology);
-    final var service = new ExportingRequestBroadcaster(client);
-
-    // when
-    service.pauseExporting(DEFAULT_PHYSICAL_TENANT_ID).join();
-
-    // then
-    for (final var partition : topology.getPartitions()) {
-      for (final var follower :
-          Optional.ofNullable(topology.getFollowersForPartition(partition)).orElse(Set.of())) {
-        verify(client).sendRequest(requestTo(partition, follower));
-      }
-      verify(client).sendRequest(requestTo(partition, topology.getLeaderForPartition(partition)));
-
-      for (final var inactive :
-          Optional.ofNullable(topology.getInactiveNodesForPartition(partition)).orElse(Set.of())) {
-        verify(client).sendRequest(requestTo(partition, inactive));
-      }
-    }
-  }
-
-  @ParameterizedTest
-  @MethodSource("invalidTopologies")
-  void shouldFailOnIncompleteTopology(final BrokerClusterState topology) {
-    // given
-    final var client = setupBrokerClient(topology);
-    final var service = new ExportingRequestBroadcaster(client);
-
-    // then
-    assertThatExceptionOfType(IncompleteTopologyException.class)
-        .isThrownBy(() -> service.pauseExporting(DEFAULT_PHYSICAL_TENANT_ID));
-  }
-
-  @ParameterizedTest
-  @MethodSource("validTopologies")
-  void shouldSucceedIfAllRequestsFinish(final BrokerClusterState topology) {
-    // given
-    final var client = setupBrokerClient(topology);
-    final var service = new ExportingRequestBroadcaster(client);
-
-    // then
-    assertThat(service.pauseExporting(DEFAULT_PHYSICAL_TENANT_ID))
-        .succeedsWithin(Duration.ofSeconds(10));
-  }
-
-  @ParameterizedTest
-  @MethodSource("validTopologies")
-  void shouldFailIfAnyRequestFails(final BrokerClusterState topology) {
-    // given
-    final var client = setupBrokerClient(topology);
-    final var service = new ExportingRequestBroadcaster(client);
-
-    // when
-    when(client.sendRequest(requestTo(1, BrokerMemberId.from(1))))
-        .thenThrow(new RuntimeException("request failed"));
-
-    // then
-    assertThatExceptionOfType(Throwable.class)
-        .isThrownBy(() -> service.pauseExporting(DEFAULT_PHYSICAL_TENANT_ID));
-  }
-
-  @ParameterizedTest
-  @MethodSource("validTopologies")
-  void shouldReportSinglePhaseWhenAllReplicasAgree(final BrokerClusterState topology) {
-    // given
-    final var client = setupBrokerClient(topology);
-    final var service = new ExportingRequestBroadcaster(client);
-    when(client.sendRequest(any())).thenAnswer(invocation -> respondWithPhase("PAUSED"));
-
-    // when
-    final var status = service.getExportingStatus(DEFAULT_PHYSICAL_TENANT_ID);
-
-    // then
-    assertThat(status).succeedsWithin(Duration.ofSeconds(10)).isEqualTo(ExportingStatus.PAUSED);
-  }
-
-  @ParameterizedTest
-  @MethodSource("validTopologies")
-  void shouldReportMixedWhenReplicasDisagree(final BrokerClusterState topology) {
-    // given a single replica that has not caught up with the pause yet
-    final var client = setupBrokerClient(topology);
-    final var service = new ExportingRequestBroadcaster(client);
-    when(client.sendRequest(any())).thenAnswer(invocation -> respondWithPhase("PAUSED"));
-    when(client.sendRequest(requestTo(1, topology.getLeaderForPartition(1))))
-        .thenAnswer(invocation -> respondWithPhase("EXPORTING"));
-
-    // when
-    final var status = service.getExportingStatus(DEFAULT_PHYSICAL_TENANT_ID);
-
-    // then
-    assertThat(status).succeedsWithin(Duration.ofSeconds(10)).isEqualTo(ExportingStatus.MIXED);
-  }
-
-  @ParameterizedTest
-  @MethodSource("validTopologies")
-  void shouldFailWhenReplicasReportAnUnknownPhase(final BrokerClusterState topology) {
-    // given every replica agrees, but on a phase this gateway does not know -- as would happen if
-    // a phase were added to the broker without updating the aggregation
-    final var client = setupBrokerClient(topology);
-    final var service = new ExportingRequestBroadcaster(client);
-    when(client.sendRequest(any())).thenAnswer(invocation -> respondWithPhase("SOME_FUTURE_PHASE"));
-
-    // when
-    final var status = service.getExportingStatus(DEFAULT_PHYSICAL_TENANT_ID);
-
-    // then it must not be silently reported as MIXED, which would read as a benign "in flight"
-    assertThat(status)
-        .failsWithin(Duration.ofSeconds(10))
-        .withThrowableOfType(ExecutionException.class)
-        .havingCause()
-        .isInstanceOf(IllegalArgumentException.class)
-        .withMessageContaining("SOME_FUTURE_PHASE");
-  }
-
-  @ParameterizedTest
-  @MethodSource("validTopologies")
-  void shouldQueryInactiveMembersForStatus(final BrokerClusterState topology) {
-    // given
-    final var client = setupBrokerClient(topology);
-    final var service = new ExportingRequestBroadcaster(client);
-    when(client.sendRequest(any())).thenAnswer(invocation -> respondWithPhase("PAUSED"));
-
-    // when
-    service.getExportingStatus(DEFAULT_PHYSICAL_TENANT_ID).join();
-
-    // then inactive members are asked too: they are only temporarily unhealthy and keep their
-    // partition data, so a pause that did not reach them is not a complete pause
-    for (final var partition : topology.getPartitions()) {
-      for (final var inactive : topology.getInactiveNodesForPartition(partition)) {
-        verify(client).sendRequest(requestTo(partition, inactive));
-      }
-    }
-  }
-
-  @ParameterizedTest
-  @MethodSource("invalidTopologies")
-  void shouldFailToReportStatusOnIncompleteTopology(final BrokerClusterState topology) {
-    // given
-    final var client = setupBrokerClient(topology);
-    final var service = new ExportingRequestBroadcaster(client);
-
-    // then
-    assertThatExceptionOfType(IncompleteTopologyException.class)
-        .isThrownBy(() -> service.getExportingStatus(DEFAULT_PHYSICAL_TENANT_ID));
-  }
-
-  private static CompletableFuture<BrokerResponse<AdminResponse>> respondWithPhase(
-      final String phase) {
-    final var response = mock(AdminResponse.class);
-    when(response.getPayload()).thenReturn(phase.getBytes(StandardCharsets.UTF_8));
-    return CompletableFuture.completedFuture(new BrokerResponse<>(response));
-  }
-
-  private BrokerClient setupBrokerClient(final BrokerClusterState topology) {
-    final var client = mock(BrokerClient.class);
-    final var topologyManager = mock(BrokerTopologyManager.class);
-
-    when(topologyManager.getTopology(DEFAULT_PHYSICAL_TENANT_ID)).thenReturn(topology);
-    when(client.getTopologyManager()).thenReturn(topologyManager);
-    when(client.sendRequest(any()))
-        .thenReturn(CompletableFuture.completedFuture(new BrokerResponse<>(new AdminResponse())));
-    return client;
-  }
-
-  public static Stream<Arguments> validTopologies() {
-    return Stream.of(
-        arguments(
-            named(
-                "Evenly distributed",
-                ofTopology(
-                    Map.of(
-                        1,
-                        List.of(
-                            BrokerMemberId.from(1), BrokerMemberId.from(2), BrokerMemberId.from(3)),
-                        2,
-                        List.of(
-                            BrokerMemberId.from(2), BrokerMemberId.from(1), BrokerMemberId.from(3)),
-                        3,
-                        List.of(
-                            BrokerMemberId.from(3),
-                            BrokerMemberId.from(1),
-                            BrokerMemberId.from(2)))))),
-        arguments(
-            named(
-                "Single broker, no replication",
-                ofTopology(
-                    Map.of(
-                        1,
-                        List.of(BrokerMemberId.from(1)),
-                        2,
-                        List.of(BrokerMemberId.from(1)),
-                        3,
-                        List.of(BrokerMemberId.from(1)))))),
-        arguments(
-            named(
-                "Multiple brokers, no replication",
-                ofTopology(
-                    Map.of(
-                        1,
-                        List.of(BrokerMemberId.from(1)),
-                        2,
-                        List.of(BrokerMemberId.from(2)),
-                        3,
-                        List.of(BrokerMemberId.from(3)))))));
-  }
-
-  public static Stream<Arguments> invalidTopologies() {
-    return Stream.of(
-        arguments(named("Partition without members", ofTopology(3, 1, 3, Map.of(1, List.of())))),
-        arguments(
-            named(
-                "Partition with missing member",
-                ofTopology(
-                    3, 1, 3, Map.of(1, List.of(BrokerMemberId.from(1), BrokerMemberId.from(2)))))),
-        arguments(named("Empty topology", ofTopology(1, 1, 1, Map.of()))));
-  }
-
-  private static BrokerClusterState ofTopology(
-      final int clusterSize,
-      final int partitionCount,
-      final int replicationFactor,
-      final Map<Integer, List<BrokerMemberId>> topology) {
-    return new BrokerClusterState() {
-      @Override
-      public boolean isInitialized() {
-        return true;
-      }
-
-      @Override
-      public int getClusterSize() {
-        return clusterSize;
-      }
-
-      @Override
-      public int getPartitionsCount() {
-        return partitionCount;
-      }
-
-      @Override
-      public int getReplicationFactor() {
-        return replicationFactor;
-      }
-
-      @Override
-      public @Nullable BrokerMemberId getLeaderForPartition(final int partition) {
-        return Optional.ofNullable(topology.get(partition))
-            .filter(brokers -> !brokers.isEmpty())
-            .map(List::getFirst)
-            .orElse(null);
-      }
-
-      @Override
-      public Set<BrokerMemberId> getFollowersForPartition(final int partition) {
-        return topology.getOrDefault(partition, List.of()).stream()
-            .skip(1)
-            .collect(Collectors.toSet());
-      }
-
-      @Override
-      public Set<BrokerMemberId> getInactiveNodesForPartition(final int partition) {
-        final var members = topology.getOrDefault(partition, List.of());
-
-        return getBrokers().stream()
-            .filter(broker -> !members.contains(broker))
-            .collect(Collectors.toSet());
-      }
-
-      @Override
-      public @Nullable BrokerMemberId getRandomBroker() {
-        throw new UnsupportedOperationException();
-      }
-
-      @Override
-      public List<Integer> getPartitions() {
-        return new ArrayList<>(topology.keySet());
-      }
-
-      @Override
-      public List<BrokerMemberId> getBrokers() {
-        return topology.values().stream().flatMap(Collection::stream).toList();
-      }
-
-      @Override
-      public @Nullable String getBrokerAddress(final @NonNull BrokerMemberId brokerId) {
-        throw new UnsupportedOperationException();
-      }
-
-      @Override
-      public @Nullable String getBrokerVersion(final @NonNull BrokerMemberId brokerId) {
-        throw new UnsupportedOperationException();
-      }
-
-      @Override
-      public PartitionHealthStatus getPartitionHealth(
-          final @NonNull BrokerMemberId brokerId, final int partition) {
-        throw new UnsupportedOperationException();
-      }
-
-      @Override
-      public State getPartitionState(final BrokerMemberId brokerId, final int partition) {
-        throw new UnsupportedOperationException();
-      }
-
-      @Override
-      public long getLastCompletedChangeId() {
-        return 0;
-      }
-
-      @Override
-      public String getClusterId() {
-        throw new UnsupportedOperationException();
-      }
-    };
-  }
-
-  private static BrokerClusterState ofTopology(final Map<Integer, List<BrokerMemberId>> topology) {
-    final var brokers = topology.values().stream().flatMap(Collection::stream).distinct().count();
-    final var partitions = topology.size();
-    final var replicationFactor =
-        topology.values().stream()
-            .map(Collection::size)
-            .max(Comparator.comparingInt(x -> x))
-            .orElseThrow();
-    return ofTopology((int) brokers, partitions, replicationFactor, topology);
-  }
-
-  record RequestMatcher(int partitionId, BrokerMemberId brokerId)
-      implements ArgumentMatcher<BrokerRequest<Void>> {
-
-    static BrokerRequest<Void> requestTo(final int partitionId, final BrokerMemberId brokerId) {
-      return argThat(new RequestMatcher(partitionId, brokerId));
-    }
-
-    @Override
-    public boolean matches(final BrokerRequest<Void> argument) {
-      return argument.getPartitionId() == partitionId
-          && argument.getBrokerId().orElseThrow().equals(brokerId);
-    }
-  }
-}
